Make Omorphia an ESM-only module
ESM is widely supported these days and continues to improve. All
intended Omorphia consumers do not use CJS export, as such it is
unnecessary and can be safely removed to speed up the build process
and remove the possibility of the dual package hazard.

Third-party consumers recommended to switch to ESM or use async import.

Recommended read:
https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c