Add support for translations using VIntl
Adds support for translations using VIntl module as a peer dependency.

Peer dependencies are dependencies that have to be provided by the
consumer, which means that `@vintl/vintl` has to be installed and set up
in Theseus and any other consumers.

Translations for Omorphia are done in Omorphia to prevent duplicate
work, and the consumer has to import the locale files like they do for
with the styles already. Those locale files have to be registered in
VIntl plugin and processed using `@vintl/unplugin`.

`@vintl/unplugin` is a Unplugin (which means it works with Rollup, Vite,
and even Webpack) that takes in locale files generated by
`@formatjs/cli` utility and transforms them into the proper ES modules.
This allows consumers to get rid of the `intl-messageformat-parser`,
which has a huge toll on the bundle size otherwise, at the cost of
inability to parse custom messages.

To give an example of how this can be approached, docs have been updated
to support changing the language used in examples. Knossos is already
doing all this automatically by using `@vintl/nuxt`.

`@formatjs/cli` allows to define messages directly and the code and then
automatically extract them into the locale file. Collocation makes it
easier to maintain the messages and see them right in the code.

Translations are done using Crowdin, so Crowdin needs to be connected to
this repository, just like it was connected to Knossos [^1]. This commit
also introduces a config that allows to do just that.

[^1]: https://github.com/modrinth/knossos/pull/738#merge-and-integration